+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Features
+
+- Add C++ exception monitor option ([#554](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/554))
+  - **Compose Multiplatform**: It is highly recommended to disable C++ monitoring (`options.enableUnhandledCppExceptionMonitoring = false`) on Apple targets. Unhandled Kotlin exceptions from CMP can be reported as generic C++ crashes (`ExceptionObjHolderImpl`) instead of useful Kotlin stack traces when C++ monitoring is enabled.
+
 ### Dependencies
 
 - Bump Java SDK from v8.40.0 to v8.41.0 ([#552](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/552))